[Android/Feat] 루틴 생성 시 조명 컨트롤 UI 구현 및 컨트롤 정보 받아오기 성공

+++ Android/domain/src/main/java/com/hogumiwarts/domain/usecase/LightUseCase.kt
@@ -11,13 +11,13 @@ import javax.inject.Inject
 class LightUseCase @Inject constructor(
     private val lightRepository: LightRepository
 ) {
-    suspend fun getLightStatus(deviceId: Long): GetLightStatusResult {
+    suspend fun getLightStatus(deviceId: Int): GetLightStatusResult {
         val data = lightRepository.getLightStatus(deviceId)
         return data
     }
 
-    suspend fun patchLightPower(deviceId: Long, activated: Boolean): PatchSwitchPowerResult {
-        val data = lightRepository.patchLightPower(deviceId =deviceId,activated = activated)
+    suspend fun patchLightPower(deviceId: Int, activated: Boolean): PatchSwitchPowerResult {
+        val data = lightRepository.patchLightPower(deviceId =deviceId, activated = activated)
         return data
     }
